Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -10,6 +10,6 @@ int main(){
     int d;
     d = (a+b+c)3;
 
-    printf("Average:%d\n",d);
+    printf("Average:%d\n",(float)d);
     return 0;
 }
